refactor(rag): split evaluation support from rag service
